syncviewmodel integration
