+++ Classes/Controllers/Settings/ARLabSettingsSyncViewController.swift
@@ -1,27 +1,55 @@
 import CoreData
+//import SyncLog
 
 class ARLabSettingsSyncViewController:UIViewController, UITableViewDelegate, UITableViewDataSource {
 
     @IBOutlet weak var lastSyncedTableView: UITableView!
+    @IBOutlet weak var tableView: UITableView!
+
+    let viewModel = ARSyncStatusViewModel.init(sync: nil);
 
     override func viewDidLoad() {
         self.lastSyncedTableView.dataSource = self
     }
 
     func tableView(tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return syncLogCount()
+        return self.viewModel.syncLogCount()
     }
 
     func tableView(tableView: UITableView, cellForRowAtIndexPath indexPath: NSIndexPath) -> UITableViewCell {
+
         let cell = tableView.dequeueReusableCellWithIdentifier("lastSyncedCell") as UITableViewCell!
-        cell.textLabel?.text = "November 17, 1992 at 5:00am"
+
+        if let dates = self.viewModel.lastSyncedStrings() as? [String] {
+            if let textLabel = cell.textLabel as UILabel! {
+                textLabel.text = dates[indexPath.row]
+                textLabel.font = UIFont.serifFontWithSize(15)
+                textLabel.textColor = UIColor.artsyHeavyGrey()
+            }
+        }
+
         return cell
     }
 
-    func syncLogCount() -> NSInteger {
-        let request:NSFetchRequest = NSFetchRequest(entityName: "SyncLog")
-        request.sortDescriptors = [NSSortDescriptor(key:"dateStarted", ascending: true)]
-        return CoreDataManager.mainManagedObjectContext().countForFetchRequest(request, error: nil)
+    @IBAction func syncButtonPressed(sender: AnyObject) {
+        self.viewModel.startSync()
+        self.tableView.reloadData()
     }
+}
 
+
+/// http://stackoverflow.com/questions/25925914/attributed-string-with-custom-fonts-in-storyboard-does-not-load-correctly
+@IBDesignable class ARAttributedLabel: UILabel {
+
+    @IBInspectable var fontSize: CGFloat = 16.0
+
+    @IBInspectable var fontFamily: String = "AGaramondPro-Regular"
+
+    override func awakeFromNib() {
+        let attrString = NSMutableAttributedString(attributedString: self.attributedText!)
+
+        attrString.addAttribute(NSFontAttributeName, value: UIFont(name: self.fontFamily, size: self.fontSize)!, range: NSMakeRange(0, attrString.length))
+
+        self.attributedText = attrString
+    }
 }
